fe: comps: moved multipage form in its folder
